Change config system to box mode and change cron check device
