+++ front/src/routes/settings/settings-system/SettingsSystemPage.jsx
@@ -1,7 +1,13 @@
 import { Text } from 'preact-i18n';
 import SettingsLayout from '../SettingsLayout';
-import cx from 'classnames';
-import Select from 'react-select';
+import SettingsSystemBatteryLevelWarning from './SettingsSystemBatteryLevelWarning';
+import SettingsSystemContainers from './SettingsSystemContainers';
+import SettingsSystemOperations from './SettingsSystemOperations';
+import SettingsSystemTimezone from './SettingsSystemTimezone';
+import SettingsSystemKeepDeviceHistory from './SettingsSystemKeepDeviceHistory';
+import SettingsSystemKeepAggregatedStates from './SettingsSystemKeepAggregatedStates';
+import SettingsSystemTimeExpiryState from './SettingsSystemTimeExpiryState';
+import SettingsSystemDatabaseCleaning from './SettingsSystemDatabaseCleaning';
 
 const SystemPage = ({ children, ...props }) => (
   <SettingsLayout>
@@ -79,276 +85,16 @@ const SystemPage = ({ children, ...props }) => (
     </div>
     <div class="row">
       <div class="col-lg-6">
-        <div class="card">
-          <h3 class="card-header">
-            <Text id="systemSettings.operations" />
-          </h3>
-
-          {props.systemInfos && props.systemInfos.new_release_available === true && (
-            <div class="card-body">
-              <div>
-                <h4>
-                  <Text id="systemSettings.newUpgradeAvailable" />
-                </h4>
-                <p>
-                  <Text id="systemSettings.newUpgradeAvailableText" />
-                </p>
-              </div>
-            </div>
-          )}
-
-          {props.systemInfos && props.systemInfos.new_release_available === false && (
-            <div class="table-responsive">
-              <table class="table table-hover table-outline table-vcenter text-nowrap card-table">
-                <tbody>
-                  <tr>
-                    <td>
-                      <Text id="systemSettings.upToDate" />
-                    </td>
-                    <td class="text-right">
-                      <span class="badge badge-success">
-                        <Text
-                          id="systemSettings.gladysVersionValue"
-                          fields={{ version: props.systemInfos.gladys_version }}
-                        />
-                      </span>
-                    </td>
-                  </tr>
-                </tbody>
-              </table>
-            </div>
-          )}
-        </div>
-
-        <div class="card">
-          <h3 class="card-header">
-            <Text id="systemSettings.configurationTitle" />
-          </h3>
-          <div class="card-body">
-            <form>
-              <label>
-                <Text id="systemSettings.timezone" />
-              </label>
-              <p>
-                <small>
-                  <Text id="systemSettings.timezoneText" />
-                </small>
-              </p>
-              <Select options={props.timezoneOptions} onChange={props.updateTimezone} value={props.selectedTimezone} />
-            </form>
-            <form class="mt-4">
-              <label>
-                <Text id="systemSettings.deviceStateRetentionTime" />
-              </label>
-              <p>
-                <small>
-                  <Text id="systemSettings.deviceStateRetentionTimeDescription" />
-                </small>
-              </p>
-              <select
-                class="form-control"
-                value={props.deviceStateHistoryInDays}
-                onChange={props.updateDeviceStateHistory}
-              >
-                <option value="7">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationOneWeek" />
-                </option>
-                <option value="30">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationOneMonth" />
-                </option>
-                <option value="90">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationThreeMonth" />
-                </option>
-                <option value="180">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationSixMonths" />
-                </option>
-                <option value="365">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationOneYear" />
-                </option>
-                <option value="730">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationTwoYears" />
-                </option>
-                <option value="-1">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.unlimited" />
-                </option>
-              </select>
-            </form>
-            <form class="mt-4">
-              <label>
-                <Text id="systemSettings.deviceAggregatesStateRetentionTime" />
-              </label>
-              <p>
-                <small>
-                  <Text id="systemSettings.deviceAggregatesStateRetentionTimeDescription" />
-                </small>
-              </p>
-              <select
-                class="form-control"
-                value={props.deviceAggregateStateHistoryInDays}
-                onChange={props.updateDeviceAggregateStateHistory}
-              >
-                <option value="7">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationOneWeek" />
-                </option>
-                <option value="30">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationOneMonth" />
-                </option>
-                <option value="90">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationThreeMonth" />
-                </option>
-                <option value="180">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationSixMonths" />
-                </option>
-                <option value="365">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationOneYear" />
-                </option>
-                <option value="730">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.durationTwoYears" />
-                </option>
-                <option value="-1">
-                  <Text id="signup.preferences.deviceStateHistoryDuration.unlimited" />
-                </option>
-              </select>
-            </form>
-            <form class="mt-4">
-              <label>
-                <Text id="systemSettings.numberOfHoursBeforeStateIsOutdated" />
-              </label>
-              <p>
-                <small>
-                  <Text id="systemSettings.numberOfHoursBeforeStateIsOutdatedDescription" />
-                </small>
-              </p>
-              <div class="input-group">
-                <input
-                  class="form-control"
-                  type="number"
-                  min="1"
-                  disabled={props.savingNumberOfHourseBeforeStateIsOutdated}
-                  value={props.numberOfHoursBeforeStateIsOutdated}
-                  onChange={props.updateNumberOfHoursBeforeStateIsOutdated}
-                />
-                <div class="input-group-append">
-                  <span class="input-group-text">
-                    <Text id="systemSettings.hours" />
-                  </span>
-                </div>
-              </div>
-            </form>
-            <form class="mt-4">
-              <label>
-                <Text id="systemSettings.vacuumDatabaseTitle" />
-              </label>
-              <p>
-                <small>
-                  <Text id="systemSettings.vacuumDatabaseDescription" />
-                </small>
-              </p>
-              <p>
-                {props.vacuumStarted && (
-                  <div class="alert alert-info">
-                    <Text id="systemSettings.vacuumDatabaseStarted" />
-                  </div>
-                )}
-                <button onClick={props.vacuumDatabase} class="btn btn-primary">
-                  <Text id="systemSettings.vacuumDatabaseButton" />
-                </button>
-              </p>
-            </form>
-          </div>
-        </div>
-        <div class="card">
-          <h3 class="card-header d-flex flex-row justify-content-between">
-            <label
-              class={cx('mb-0', {
-                'text-muted': !props.batteryLevelUnderWarningEnabled
-              })}
-            >
-              <Text id="systemSettings.batteryLevel" />
-            </label>
-            <label className="custom-switch">
-              <input
-                type="checkbox"
-                name="active"
-                value="1"
-                className="custom-switch-input"
-                checked={props.batteryLevelUnderWarningEnabled}
-                onClick={props.updateBatteryLevelUnderWarningEnabled}
-              />
-              <span class="custom-switch-indicator" />
-            </label>
-          </h3>
-          <div class="card-body">
-            <form className="">
-              <p
-                class={cx({
-                  'text-muted': !props.batteryLevelUnderWarningEnabled
-                })}
-              >
-                <Text id="systemSettings.batteryLevelDescription" />
-              </p>
-              <div class="input-group">
-                <input
-                  className="form-control"
-                  type="number"
-                  min="1"
-                  max="100"
-                  disabled={props.savingBatteryLevelUnderWarning || !props.batteryLevelUnderWarningEnabled}
-                  value={props.batteryLevelUnderWarningThreshold}
-                  onChange={props.updateBatteryLevelUnderWarningThreshold}
-                />
-                <div class="input-group-append">
-                  <span class="input-group-text">
-                    <Text id="global.percent" />
-                  </span>
-                </div>
-              </div>
-            </form>
-          </div>
-        </div>
+        <SettingsSystemOperations />
+        <SettingsSystemKeepDeviceHistory />
+        <SettingsSystemKeepAggregatedStates />
+        <SettingsSystemTimeExpiryState />
       </div>
       <div class="col-lg-6">
-        <div class="card">
-          <h3 class="card-header">
-            <Text id="systemSettings.containers" />
-          </h3>
-          <div class="table-responsive" style={{ maxHeight: '200px' }}>
-            <table class="table table-hover table-outline table-vcenter text-nowrap card-table">
-              <thead>
-                <tr>
-                  <th>
-                    <Text id="systemSettings.containerName" />
-                  </th>
-                  <th>
-                    <Text id="systemSettings.containerCreated" />
-                  </th>
-                  <th>
-                    <Text id="systemSettings.containerStatus" />
-                  </th>
-                </tr>
-              </thead>
-              <tbody>
-                {props.systemContainers &&
-                  props.systemContainers.map(container => (
-                    <tr>
-                      <td>{container.name}</td>
-                      <td>{container.created_at_formatted}</td>
-                      <td>
-                        <span
-                          class={cx('badge', {
-                            'badge-success': container.state === 'running',
-                            'badge-warning': container.state !== 'running'
-                          })}
-                        >
-                          <Text id={`systemSettings.containerState.${container.state}`} />
-                        </span>
-                      </td>
-                    </tr>
-                  ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
+        <SettingsSystemTimezone />
+        <SettingsSystemBatteryLevelWarning />
+        <SettingsSystemDatabaseCleaning />
+        <SettingsSystemContainers />
       </div>
     </div>
   </SettingsLayout>
